change price to Double

+++ java-backend/src/main/java/com/example/cardapio/product/ProductResponseDTO.java
@@ -1,6 +1,6 @@
 package com.example.cardapio.product;
 
-public record ProductResponseDTO(Long id, String title, String image, Integer price) {
+public record ProductResponseDTO(Long id, String title, String image, Double price) {
     public ProductResponseDTO(Product product) {
         this(product.getId(), product.getTitle(), product.getImage(), product.getPrice());
     }
